Add automatic flow, interpret target and action variables
An automatic flow has been added. This means that when there no
run arguments indicating what to do, hdlmake tries to guess
the action on its own basing on the manifest.

Manifest's target indicates what is the synthesis target
(xilinx/altera). This influences what kind of project for synthesis
is made (qpf, xise).

Manifest's action indicates, if the manifest is used for synthesis,
simulation or cannot be treated as a top manifest.