Fix make_dependency_sorted_list ordering function.
Previous implelmentaiton only worked for a single level tree of dependencies.
This implementations recursively determines the number of levels below a dep_file and uses this value as the key to sort by. The 'dependendency level is cached in the dep_file object to reduce the number of recursive calls.
Will not handle circular dependencies well.